0.7.0 - refactored for clarity, new rollup config, updated rating-scales
