+++ response-items/ellipse-component/package.json
@@ -1,14 +1,12 @@
 {
   "name": "@decsys/ellipse-component",
-  "bundle": "EllipseScale",
-  "version": "0.6.0",
+  "componentName": "Ellipse",
+  "version": "0.7.0",
   "description": "The Ellipse Scale component for the DECSYS Survey Platform",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "rollup": "rollup -c",
-    "esm": "node ./build/esm-export",
-    "build": "npm run rollup && npm run esm",
+    "build": "npm run rollup",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
@@ -33,11 +31,12 @@
     "rollup-plugin-jsx": "^1.0.3",
     "rollup-plugin-node-resolve": "^4.0.1",
     "rollup-plugin-replace": "^2.1.1",
+    "rollup-plugin-terser": "^4.0.4",
     "styled-components": "^4.1.3"
   },
   "dependencies": {
-    "@decsys/param-types": "^0.3.0",
-    "@decsys/rating-scales": "^0.6.5",
+    "@decsys/param-types": "^0.3.1",
+    "@decsys/rating-scales": "^0.8.0",
     "styled-icons": "^7.6.0"
   },
   "peerDependencies": {
