+++ response-items/ellipse-component/src/metadata.js
@@ -1,0 +1,3 @@
+import { CircleNotch } from "styled-icons/fa-solid/CircleNotch";
+
+export const icon = <CircleNotch />;
